requirements.txt
django==4.1.2
python-decouple==3.6
djangorestframework==3.14.0
django-rest-swagger==2.2.0
pyTelegramBotAPI==4.7.0
